+++ odoo/local-src/specific_purchase/__manifest__.py
@@ -5,14 +5,11 @@
 {
     "name": "RocTool specific Purchase Module",
     "version": "10.0.1.0.0",
-    "depends": [
-        'purchase',
-        'hr',
-    ],
+    "depends": ['purchase'],
     "author": "Camptocamp,Odoo Community Association (OCA)",
     "website": "http://www.camptocamp.com",
     "license": "GPL-3 or any later version",
-    "category": "Sale",
+    "category": "Purchase",
     "data": [
         # security
         'security/ir.model.access.csv',