+++ odoo/local-src/specific_purchase/models/purchase.py
@@ -3,7 +3,7 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import models
+from odoo import models, api, exceptions, _
 
 
 class PurchaseOrder(models.Model):
@@ -15,3 +15,12 @@ class PurchaseOrder(models.Model):
                 super(PurchaseOrder, self).write(vals)
         else:
             super(PurchaseOrder, self).write(vals)
+
+    @api.multi
+    def button_approve(self, force=False):
+        for order in self:
+            for line in order.order_line:
+                if not line.account_analytic_id:
+                    raise exceptions.UserError(
+                        _('An Analytic Account is required for Validation!'))
+        return super(PurchaseOrder, self).button_approve(force=force)